requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
websockets==12.0
pyyaml>=6.0.1
pillow>=10.0.0
